Added Browser.close() for android
I'm new to ionic and capacitor and I had a problem I couldn't solve: Browser cannot be closed on android. I found out that you can't close CustomTabs on android. After searching for a while, I found out an optional workaround for that problem (based on an this answer on StackOverflow: https://stackoverflow.com/a/41596629/5976838).
The workaround works perfectly for me so I uploaded it here.