Use the default referrer for launching CustomTab on Android
Browser plugin has a bug setting the referrer of an invalid form:

1. Wrong scheme: should be android-app but set to 2
2. Missing colon

The bug results a referrer such as 2//com.package.app while the
intention is to build something like android-app://com.package.app.

This patch removes the line with the bug. Android fills out the
referrer with the form the code intended by default.