fix(ios): iOS panorama photos selected through CameraPlugin are corrupted
The bug was due to an out-of-memory issue. The `self.draw` call kept the image in memory after usage. When transforming the image to a jpeg data, there wasn't enough memory available to perform the operation, so a `nil` data object was being returned, resulting in an empty image.
Using `UIGraphicsxxxImageContext` fixes the issue.